feat: 리뷰 수정, 삭제 시 접근 권한 검사 추가

+++ src/main/java/com/kt/service/ReviewServiceImpl.java
@@ -1,6 +1,5 @@
 package com.kt.service;
 
-import java.util.List;
 import java.util.UUID;
 
 import org.springframework.data.domain.Page;
@@ -9,11 +8,13 @@ import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
 import com.kt.constant.OrderStatus;
+import com.kt.constant.UserRole;
 import com.kt.constant.message.ErrorCode;
 import com.kt.constant.searchtype.ProductSearchType;
 import com.kt.domain.dto.response.ReviewResponse;
 import com.kt.domain.entity.OrderProductEntity;
 import com.kt.domain.entity.ReviewEntity;
+import com.kt.domain.entity.UserEntity;
 import com.kt.exception.CustomException;
 import com.kt.repository.orderproduct.OrderProductRepository;
 import com.kt.repository.review.ReviewRepository;
@@ -28,6 +29,21 @@ public class ReviewServiceImpl implements ReviewService {
 	private final ReviewRepository reviewRepository;
 	private final OrderProductRepository orderProductRepository;
 
+	private static boolean hasReviewUpdatePermission(String email, ReviewEntity review) {
+		UserEntity orderBy = review
+			.getOrderProduct()
+			.getOrder()
+			.getOrderBy();
+
+		if (orderBy.getRole() == UserRole.ADMIN) {
+			return true;
+		}
+		if (email.equals(orderBy.getEmail())) {
+			return true;
+		}
+		return false;
+	}
+
 	@Override
 	public void create(UUID orderProductId, String content) {
 		OrderProductEntity orderProduct = orderProductRepository.findByIdOrThrow(orderProductId);
@@ -40,14 +56,25 @@ public class ReviewServiceImpl implements ReviewService {
 	}
 
 	@Override
-	public void update(UUID reviewId, String content) {
+	public void update(
+		String email,
+		UUID reviewId,
+		String content
+	) {
 		ReviewEntity review = reviewRepository.findByIdOrThrow(reviewId);
+		if (!hasReviewUpdatePermission(email, review))
+			throw new CustomException(ErrorCode.REVIEW_ACCESS_NOT_ALLOWED);
 		review.update(content);
 	}
 
 	@Override
-	public void delete(UUID reviewId) {
+	public void delete(
+		String email,
+		UUID reviewId
+	) {
 		ReviewEntity review = reviewRepository.findByIdOrThrow(reviewId);
+		if (!hasReviewUpdatePermission(email, review))
+			throw new CustomException(ErrorCode.REVIEW_ACCESS_NOT_ALLOWED);
 		review.delete();
 	}
 
@@ -61,26 +88,20 @@ public class ReviewServiceImpl implements ReviewService {
 	}
 
 	@Override
-	public List<ReviewResponse.Search> getReviewByProductId(UUID productId) {
-		List<OrderProductEntity> orderProducts = orderProductRepository.findAllByProductId(productId);
-
-		return orderProducts.stream().map(orderProduct -> {
-			ReviewEntity reviewEntity = reviewRepository.findByOrderProductIdOrThrow(orderProduct.getId());
-			return new ReviewResponse.Search(
-				reviewEntity.getId(),
-				reviewEntity.getContent()
-			);
-		}).toList();
-	}
-
-	@Override
-	public Page<ReviewResponse.Search> getReviewByProductId(UUID productId, Pageable pageable){
+	public Page<ReviewResponse.Search> getReviewByProductId(
+		UUID productId,
+		Pageable pageable
+	) {
 		return reviewRepository.searchReviewsByProductId(pageable, productId);
 	}
 
 	@Override
-	public Page<ReviewResponse.Search> getReviewsByAdmin(Pageable pageable, String keyword, ProductSearchType type){
-		return reviewRepository.searchReviews(pageable,keyword,type);
+	public Page<ReviewResponse.Search> getReviewsByAdmin(
+		Pageable pageable,
+		String keyword,
+		ProductSearchType type
+	) {
+		return reviewRepository.searchReviews(pageable, keyword, type);
 	}
 
 	@Override
